move all exception handlers to one folder
As there some of them, looks it cleaner to have in own folder about.
