+++ src/CMakeLists.txt.orig
@@ -45,4 +45,4 @@ set(HEADERS align.h
 
 add_library(code_src OBJECT ${SOURCES} ${HEADERS})
 target_link_libraries(code_src ${DEPLIBS})
-target_include_directories(code_src PRIVATE ${PROJECT_SOURCE_DIR}/src)
+
