+++ src/channelsettings.cpp
@@ -24,7 +24,7 @@
 
 #include "dabmuxscanner.h"
 #include "hdmuxscanner.h"
-#include "string_exception.h"
+#include "exception_control/string_exception.h"
 #include "utils/value_size_defines.h"
 
 #include <assert.h>
