+++ src/database.cpp
@@ -23,7 +23,7 @@
 #include "database.h"
 
 #include "dbtypes.h"
-#include "sqlite_exception.h"
+#include "exception_control/sqlite_exception.h"
 #include "utils/value_size_defines.h"
 
 #include <stdexcept>
