+++ src/exception_control/libusb_exception.cpp
