+++ src/exception_control/sqlite_exception.cpp
