+++ src/exception_control/sqlite_exception.h
