+++ src/exception_control/win32_exception.h
