+++ src/hdstream.cpp
@@ -24,7 +24,7 @@
 
 #include "align.h"
 #include "id3v2tag.h"
-#include "string_exception.h"
+#include "exception_control/string_exception.h"
 #include "utils/value_size_defines.h"
 
 #include <algorithm>
