+++ src/signalmeter.cpp
@@ -23,7 +23,7 @@
 #include "signalmeter.h"
 
 #include "align.h"
-#include "string_exception.h"
+#include "exception_control/string_exception.h"
 #include "utils/value_size_defines.h"
 
 #include <assert.h>
