+++ src/usbdevice.cpp
@@ -22,8 +22,8 @@
 
 #include "usbdevice.h"
 
-#include "libusb_exception.h"
-#include "string_exception.h"
+#include "exception_control/libusb_exception.h"
+#include "exception_control/string_exception.h"
 #include "utils/value_size_defines.h"
 
 #include <assert.h>
