+++ src/wxstream.cpp
@@ -23,7 +23,7 @@
 #include "wxstream.h"
 
 #include "align.h"
-#include "string_exception.h"
+#include "exception_control/string_exception.h"
 #include "utils/value_size_defines.h"
 
 #include <algorithm>
